+++ components/bpmn-q/modeler-component/extensions/opentosca/modeling/properties-provider/Connector.js
@@ -13,7 +13,7 @@ import { SelectEntry } from "@bpmn-io/properties-panel";
 import React from "@bpmn-io/properties-panel/preact/compat";
 import { useService } from "bpmn-js-properties-panel";
 import { getModeler } from "../../../../editor/ModelerHandler";
-import { HiddenSelectEntry } from "../../../../editor/popup/HiddenTextFieldEntry";
+import { HiddenSelectEntry } from "../../../../editor/popup/HiddenFieldEntry";
 import { getExtensionElementsList } from "../../../../editor/util/camunda-utils/ExtensionElementsUtil";
 import { getInputOutput } from "../../../../editor/util/camunda-utils/InputOutputUtil";
